Crecion de cursos y lecciones completo
Se prueden crear cursos y lecciones con todas las funciones, insertar, actualizar y eliminar

+++ php/crearCursos/manejoArchivos.php
@@ -5,16 +5,24 @@ define("DIRECTORIO_BORRADO",1);
 define("ERROR_DEL_DIRECTORIO",2);
 
 //Con esta funcion subimos archivos
-function subirArchivo($archivo, $directorio)
-{ 
-	$resultado=false;
+function subirArchivo($archivo, $directorio,$nuevoNombre)
+{
    //movemos el archivo al valor a la ruta especificada
-    if(move_uploaded_file($archivo["tmp_name"], $directorio."/".$archivo["name"]))
+    if(move_uploaded_file($archivo["tmp_name"], $directorio."/".$nuevoNombre))
 	{
-			$resultado=true;
+			return true;
 	}
-	return $resultado;
-		
+    else
+        return false;
+}
+
+function nuevoNombre($archivo,$codigo)
+{
+//Extraigo el nombre y la extencion del archivo y luego selecciono obtengo la extension con la funcion end
+    $nombre = explode(".", $archivo["name"]);
+    $extension=end($nombre);
+    $newName=$codigo."_".time().".".$extension;
+      return $newName;//Creamos el nuevo nombre del archivo
 }
 
 //con esta funcion eliminamos los archivos y directorios
@@ -23,18 +31,30 @@ function eliminarArchivo($url)
     //Retorna 1 si se realizo con exito
     //Retorna 2 si el parametor pasado no es un archivo
     //Retorna 0 si no se pudo borrar el archivo
-    if(is_file($url)){
-        if(unlink($url))
-        {
-            return 1;
+    //Retorna 3 si el arhivo no existe
+    if(file_exists($url))
+    {
+        if(is_file($url)){
+            if(unlink($url))
+            {
+                return 1;
+            }
+            else
+                return 0;
         }
         else
-            return 0;
+            return 2;
     }
     else
-        return 2;
+        return 3;
+
 }
 
+function renombrar($archivo,$nuevo_nombre)
+{
+
+
+}
 //Con esta funcion eliminamos un directorio
 // Devuelve 0 si lo que se paso por la url no es un direcotorio
 //Retorna 1 si el directorio fue borrado exitosamente
